+++ CHANGELOG.md
@@ -39,6 +39,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [3.4.0] - 2022-01-27
 ### Added
+- Google Tag Manager container ID. [RW-136](https://vizzuality.atlassian.net/browse/RW-136)
 - support for attributions in maps. [RW-77](https://vizzuality.atlassian.net/browse/RW-77)
 - interception to handle requests from Global Fishing Watch. [RW-117](https://vizzuality.atlassian.net/browse/RW-117)
 - 'FeatureService' provider for Layer Manager. [RW-114](https://vizzuality.atlassian.net/browse/RW-114)